+++ internal/freelist/hashmap.go
@@ -1,4 +1,4 @@
-package bbolt
+package freelist
 
 import (
 	"fmt"
@@ -8,26 +8,57 @@ import (
 	"go.etcd.io/bbolt/internal/common"
 )
 
-// hashmapFreeCount returns count of free pages(hashmap version)
-func (f *freelist) hashmapFreeCount() int {
-	common.Verify(func() {
-		expectedFreePageCount := f.hashmapFreeCountSlow()
-		common.Assert(int(f.freePagesCount) == expectedFreePageCount,
-			"freePagesCount (%d) is out of sync with free pages map (%d)", f.freePagesCount, expectedFreePageCount)
-	})
-	return int(f.freePagesCount)
+// pidSet holds the set of starting pgids which have the same span size
+type pidSet map[common.Pgid]struct{}
+
+type hashMap struct {
+	*shared
+
+	freePagesCount uint64                 // count of free pages(hashmap version)
+	freemaps       map[uint64]pidSet      // key is the size of continuous pages(span), value is a set which contains the starting pgids of same size
+	forwardMap     map[common.Pgid]uint64 // key is start pgid, value is its span size
+	backwardMap    map[common.Pgid]uint64 // key is end pgid, value is its span size
 }
 
-func (f *freelist) hashmapFreeCountSlow() int {
-	count := 0
-	for _, size := range f.forwardMap {
-		count += int(size)
+func (f *hashMap) Init(pgids common.Pgids) {
+	if len(pgids) == 0 {
+		return
 	}
-	return count
+
+	size := uint64(1)
+	start := pgids[0]
+	// reset the counter when freelist init
+	f.freePagesCount = 0
+
+	if !sort.SliceIsSorted([]common.Pgid(pgids), func(i, j int) bool { return pgids[i] < pgids[j] }) {
+		panic("pgids not sorted")
+	}
+
+	f.freemaps = make(map[uint64]pidSet)
+	f.forwardMap = make(map[common.Pgid]uint64)
+	f.backwardMap = make(map[common.Pgid]uint64)
+
+	for i := 1; i < len(pgids); i++ {
+		// continuous page
+		if pgids[i] == pgids[i-1]+1 {
+			size++
+		} else {
+			f.addSpan(start, size)
+
+			size = 1
+			start = pgids[i]
+		}
+	}
+
+	// init the tail
+	if size != 0 && start != 0 {
+		f.addSpan(start, size)
+	}
+
+	f.reindex()
 }
 
-// hashmapAllocate serves the same purpose as arrayAllocate, but use hashmap as backend
-func (f *freelist) hashmapAllocate(txid common.Txid, n int) common.Pgid {
+func (f *hashMap) Allocate(txid common.Txid, n int) common.Pgid {
 	if n == 0 {
 		return 0
 	}
@@ -74,17 +105,17 @@ func (f *freelist) hashmapAllocate(txid common.Txid, n int) common.Pgid {
 	return 0
 }
 
-// hashmapReadIDs reads pgids as input an initial the freelist(hashmap version)
-func (f *freelist) hashmapReadIDs(pgids []common.Pgid) {
-	f.init(pgids)
-
-	// Rebuild the page cache.
-	f.reindex()
+func (f *hashMap) FreeCount() int {
+	common.Verify(func() {
+		expectedFreePageCount := f.hashmapFreeCountSlow()
+		common.Assert(int(f.freePagesCount) == expectedFreePageCount,
+			"freePagesCount (%d) is out of sync with free pages map (%d)", f.freePagesCount, expectedFreePageCount)
+	})
+	return int(f.freePagesCount)
 }
 
-// hashmapGetFreePageIDs returns the sorted free page ids
-func (f *freelist) hashmapGetFreePageIDs() []common.Pgid {
-	count := f.free_count()
+func (f *hashMap) freePageIds() common.Pgids {
+	count := f.FreeCount()
 	if count == 0 {
 		return nil
 	}
@@ -108,8 +139,36 @@ func (f *freelist) hashmapGetFreePageIDs() []common.Pgid {
 	return m
 }
 
-// hashmapMergeSpans try to merge list of pages(represented by pgids) with existing spans
-func (f *freelist) hashmapMergeSpans(ids common.Pgids) {
+func (f *hashMap) hashmapFreeCountSlow() int {
+	count := 0
+	for _, size := range f.forwardMap {
+		count += int(size)
+	}
+	return count
+}
+
+func (f *hashMap) addSpan(start common.Pgid, size uint64) {
+	f.backwardMap[start-1+common.Pgid(size)] = size
+	f.forwardMap[start] = size
+	if _, ok := f.freemaps[size]; !ok {
+		f.freemaps[size] = make(map[common.Pgid]struct{})
+	}
+
+	f.freemaps[size][start] = struct{}{}
+	f.freePagesCount += size
+}
+
+func (f *hashMap) delSpan(start common.Pgid, size uint64) {
+	delete(f.forwardMap, start)
+	delete(f.backwardMap, start+common.Pgid(size-1))
+	delete(f.freemaps[size], start)
+	if len(f.freemaps[size]) == 0 {
+		delete(f.freemaps, size)
+	}
+	f.freePagesCount -= size
+}
+
+func (f *hashMap) mergeSpans(ids common.Pgids) {
 	common.Verify(func() {
 		ids1Freemap := f.idsFromFreemaps()
 		ids2Forward := f.idsFromForwardMap()
@@ -144,7 +203,7 @@ func (f *freelist) hashmapMergeSpans(ids common.Pgids) {
 }
 
 // mergeWithExistingSpan merges pid to the existing free spans, try to merge it backward and forward
-func (f *freelist) mergeWithExistingSpan(pid common.Pgid) {
+func (f *hashMap) mergeWithExistingSpan(pid common.Pgid) {
 	prev := pid - 1
 	next := pid + 1
 
@@ -171,68 +230,9 @@ func (f *freelist) mergeWithExistingSpan(pid common.Pgid) {
 	f.addSpan(newStart, newSize)
 }
 
-func (f *freelist) addSpan(start common.Pgid, size uint64) {
-	f.backwardMap[start-1+common.Pgid(size)] = size
-	f.forwardMap[start] = size
-	if _, ok := f.freemaps[size]; !ok {
-		f.freemaps[size] = make(map[common.Pgid]struct{})
-	}
-
-	f.freemaps[size][start] = struct{}{}
-	f.freePagesCount += size
-}
-
-func (f *freelist) delSpan(start common.Pgid, size uint64) {
-	delete(f.forwardMap, start)
-	delete(f.backwardMap, start+common.Pgid(size-1))
-	delete(f.freemaps[size], start)
-	if len(f.freemaps[size]) == 0 {
-		delete(f.freemaps, size)
-	}
-	f.freePagesCount -= size
-}
-
-// initial from pgids using when use hashmap version
-// pgids must be sorted
-func (f *freelist) init(pgids []common.Pgid) {
-	if len(pgids) == 0 {
-		return
-	}
-
-	size := uint64(1)
-	start := pgids[0]
-	// reset the counter when freelist init
-	f.freePagesCount = 0
-
-	if !sort.SliceIsSorted([]common.Pgid(pgids), func(i, j int) bool { return pgids[i] < pgids[j] }) {
-		panic("pgids not sorted")
-	}
-
-	f.freemaps = make(map[uint64]pidSet)
-	f.forwardMap = make(map[common.Pgid]uint64)
-	f.backwardMap = make(map[common.Pgid]uint64)
-
-	for i := 1; i < len(pgids); i++ {
-		// continuous page
-		if pgids[i] == pgids[i-1]+1 {
-			size++
-		} else {
-			f.addSpan(start, size)
-
-			size = 1
-			start = pgids[i]
-		}
-	}
-
-	// init the tail
-	if size != 0 && start != 0 {
-		f.addSpan(start, size)
-	}
-}
-
 // idsFromFreemaps get all free page IDs from f.freemaps.
 // used by test only.
-func (f *freelist) idsFromFreemaps() map[common.Pgid]struct{} {
+func (f *hashMap) idsFromFreemaps() map[common.Pgid]struct{} {
 	ids := make(map[common.Pgid]struct{})
 	for size, idSet := range f.freemaps {
 		for start := range idSet {
@@ -250,7 +250,7 @@ func (f *freelist) idsFromFreemaps() map[common.Pgid]struct{} {
 
 // idsFromForwardMap get all free page IDs from f.forwardMap.
 // used by test only.
-func (f *freelist) idsFromForwardMap() map[common.Pgid]struct{} {
+func (f *hashMap) idsFromForwardMap() map[common.Pgid]struct{} {
 	ids := make(map[common.Pgid]struct{})
 	for start, size := range f.forwardMap {
 		for i := 0; i < int(size); i++ {
@@ -266,7 +266,7 @@ func (f *freelist) idsFromForwardMap() map[common.Pgid]struct{} {
 
 // idsFromBackwardMap get all free page IDs from f.backwardMap.
 // used by test only.
-func (f *freelist) idsFromBackwardMap() map[common.Pgid]struct{} {
+func (f *hashMap) idsFromBackwardMap() map[common.Pgid]struct{} {
 	ids := make(map[common.Pgid]struct{})
 	for end, size := range f.backwardMap {
 		for i := 0; i < int(size); i++ {
@@ -278,4 +278,15 @@ func (f *freelist) idsFromBackwardMap() map[common.Pgid]struct{} {
 		}
 	}
 	return ids
+}
+
+func NewHashMapFreelist() Interface {
+	hm := &hashMap{
+		shared:      newShared(),
+		freemaps:    make(map[uint64]pidSet),
+		forwardMap:  make(map[common.Pgid]uint64),
+		backwardMap: make(map[common.Pgid]uint64),
+	}
+	hm.Interface = hm
+	return hm
 }
